My first c programm work11

+++ 0x00-hello_world/6-size.c
@@ -10,10 +10,10 @@ int b;
 long int c;
 long long int d;
 float f;
-printf("Size of a char: %lu byte(s)\n", (unsigned)sizeof(a));
-printf("Size of an int: %lu byte(s)\n", (unsigned)sizeof(b));
-printf("Size of a long int: %lu byte(s)\n", (unsigned)sizeof(c));
-printf("Size of a long long int: %lu byte(s)\n", (unsigned)sizeof(d));
-printf("Size of a float: %lu byte(s)\n", (unsigned)sizeof(f));
+printf("Size of a char: %d byte(s)\n", sizeof(a));
+printf("Size of an int: %d byte(s)\n", sizeof(b));
+printf("Size of a long int: %d byte(s)\n", sizeof(c));
+printf("Size of a long long int: %d byte(s)\n", sizeof(d));
+printf("Size of a float: %d byte(s)\n", sizeof(f));
 return (0);
 }
